refactor: 일부 이미지 리펙토링
