+++ client/src/components/common/ProfilePhoto.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-import defaultProfile from 'images/default-profile.jpg';
 
 import { ProfilePhotoProps } from 'utils/types';
 import palette from 'theme/palette';
+import { defaultProfile } from 'images';
 
 const StyledProfilePhoto = styled.img<ProfilePhotoProps>`
   width: ${(props) => props.size || '65px'};
