+++ client/src/images/index.tsx
@@ -1,0 +1,6 @@
+import mainLogo from 'images/main-logo.png';
+import defaultProfile from 'images/default-profile.jpg';
+import defaultCover from 'images/default-profile.jpg';
+import iconSearch from 'images/icon-search.svg';
+
+export { mainLogo, defaultProfile, defaultCover, iconSearch };
